ai collision

+++ src/GymnasieArbete/Game.java
@@ -144,9 +144,9 @@ public class Game extends Canvas implements Runnable {
 		
 		if (state == STATE.GAME) {
 			screen.clear();
-			int xScroll = player.getX() - screen.width / 2;
-			int yScroll = player.getY() - screen.height / 2;
-			level.render(xScroll, yScroll, screen);
+			double xScroll = player.getX() - screen.width / 2;
+			double yScroll = player.getY() - screen.height / 2;
+			level.render((int) xScroll, (int) yScroll, screen);
 			//screen.renderSheet(40, 40, SpriteSheet.player, false);
 			
 			Sprite sprite = new Sprite(width-40, 20, 0x0);
