+++ src/GymnasieArbete/entities/projectile/PistolProjectile.java
@@ -5,10 +5,10 @@ import GymnasieArbete.graphics.Screen;
 import GymnasieArbete.graphics.Sprite;
 
 public class PistolProjectile extends Projectile {
-	
+
 	public static final int FIRE_RATE = 20;
 
-	public PistolProjectile(int x, int y, double dir) {
+	public PistolProjectile(double x, double y, double dir) {
 		super(x, y, dir);
 		range = 150;
 		speed = 2;
@@ -19,19 +19,19 @@ public class PistolProjectile extends Projectile {
 	}
 
 	public void update() {
-		if(level.tileCollision((int) (x + nx), (int) (y + ny), 2, 7, 7)) {
+		if (level.tileCollision((int) (x + nx), (int) (y + ny), 2, 7, 7)) {
 			level.add(new ParticleSpawner((int) x + 5, (int) y + 6, 40, 20, level));
 			remove();
 		} else
-		move();
+			move();
 	}
 
 	protected void move() {
-			x += nx;
-			y += ny;
+		x += nx;
+		y += ny;
 		if (distance() > range) remove();
 	}
-	
+
 	private double distance() {
 		double dist = 0;
 		dist = Math.sqrt(Math.abs((xOrigin - x) * (xOrigin - x) + (yOrigin - y) * (yOrigin - y)));
